Move the RunSetup function to shared
In order to share the same code for installation, migration and upgrade
the RunSetup() function needs to move to the mgradm shared utils module.
